Refactor the CMake project structure
Prior to further work and additional files added it would be benifitial
to update the build system to use modern CMake practices.

I've updated the build system with the addition of a few more dedicated
CMakeLists.txt files and also an install config. The FindTBB.cmake file
was no longer needed so this has been removed.