Add a .clang-format config file
We can use .clang-format to make sure the code base stayes consistent
between files and developers. This can even be factored into the CI and
prevent a merge if any given file does not conform.

I've added a .clang-format file and loaded it with the default style
from LLVM. This can be changed and customised at a later date.