Increase implicit maxT value for AS algorithm
The current optimisation process is a simple hill walking algorithm, the
same as simulated annealing with a maxT value of 0. Global optima can be
better found with positive maxT which allows increases in energy.

I've added a small adjustment that makes use of a new hyper-parameter
maxT. This should allow for more optimal results.